requirements.txt
pandas
requests
langchain-openai
langchain-core
pydantic
streamlit
python-dotenv

